update tests that use GetTeamByTeamId
