+++ SocialCareCaseViewerApi.Tests/V1/UseCase/WorkersUseCaseTests.cs
@@ -7,7 +7,6 @@ using SocialCareCaseViewerApi.Tests.V1.Helpers;
 using SocialCareCaseViewerApi.V1.Boundary.Requests;
 using SocialCareCaseViewerApi.V1.Boundary.Response;
 using SocialCareCaseViewerApi.V1.Factories;
-using SocialCareCaseViewerApi.V1.Gateways;
 using SocialCareCaseViewerApi.V1.Gateways.Interfaces;
 using SocialCareCaseViewerApi.V1.Infrastructure;
 using SocialCareCaseViewerApi.V1.UseCase;
@@ -21,6 +20,7 @@ namespace SocialCareCaseViewerApi.Tests.V1.UseCase
     {
         private Mock<IDatabaseGateway> _mockDatabaseGateway = null!;
         private Mock<IWorkerGateway> _mockWorkerGateway = null!;
+        private Mock<ITeamGateway> _mockTeamGateway = null!;
         private IWorkersUseCase _workersUseCase = null!;
 
         [SetUp]
@@ -28,7 +28,8 @@ namespace SocialCareCaseViewerApi.Tests.V1.UseCase
         {
             _mockDatabaseGateway = new Mock<IDatabaseGateway>();
             _mockWorkerGateway = new Mock<IWorkerGateway>();
-            _workersUseCase = new WorkersUseCase(_mockDatabaseGateway.Object, _mockWorkerGateway.Object);
+            _mockTeamGateway = new Mock<ITeamGateway>();
+            _workersUseCase = new WorkersUseCase(_mockDatabaseGateway.Object, _mockWorkerGateway.Object, _mockTeamGateway.Object);
         }
 
         [Test]
@@ -44,7 +45,7 @@ namespace SocialCareCaseViewerApi.Tests.V1.UseCase
 
             _mockWorkerGateway.Verify(x => x.GetWorkerByWorkerId(fakeWorker.Id), Times.Once);
             _mockDatabaseGateway.Verify(x => x.GetWorkerByEmail(fakeWorker.Email), Times.Once);
-            _mockDatabaseGateway.Verify(x => x.GetTeamByTeamId(teamId), Times.Once);
+            _mockTeamGateway.Verify(x => x.GetTeamByTeamId(teamId), Times.Once);
         }
 
         [Test]
@@ -125,7 +126,7 @@ namespace SocialCareCaseViewerApi.Tests.V1.UseCase
             {
                 TeamId = fakeTeamId
             };
-            _mockDatabaseGateway.Setup(x => x.GetTeamByTeamId(fakeTeamId)).Returns(fakeTeam);
+            _mockTeamGateway.Setup(x => x.GetTeamByTeamId(fakeTeamId)).Returns(fakeTeam);
 
             var result = _workersUseCase.ExecuteGet(request);
             var worker = result.First();
@@ -163,7 +164,7 @@ namespace SocialCareCaseViewerApi.Tests.V1.UseCase
             };
             _mockWorkerGateway.Setup(x => x.GetWorkerByWorkerId(fakeWorker.Id)).Returns(fakeWorkerDomain);
             _mockDatabaseGateway.Setup(x => x.GetWorkerByEmail(fakeWorker.Email)).Returns(fakeWorker);
-            _mockDatabaseGateway.Setup(x => x.GetTeamByTeamId(fakeTeamId)).Returns(fakeTeam);
+            _mockTeamGateway.Setup(x => x.GetTeamByTeamId(fakeTeamId)).Returns(fakeTeam);
 
             var result = _workersUseCase.ExecuteGet(request);
             result.Count.Should().Be(1);
@@ -211,7 +212,7 @@ namespace SocialCareCaseViewerApi.Tests.V1.UseCase
 
             _mockWorkerGateway.Setup(x => x.GetWorkerByWorkerId(fakeWorkerId.Id)).Returns(fakeWorkerId);
             _mockDatabaseGateway.Setup(x => x.GetWorkerByEmail(fakeWorkerEmail.Email)).Returns(fakeWorkerEmail);
-            _mockDatabaseGateway.Setup(x => x.GetTeamByTeamId(fakeTeamId)).Returns(fakeTeam);
+            _mockTeamGateway.Setup(x => x.GetTeamByTeamId(fakeTeamId)).Returns(fakeTeam);
 
             var request = new GetWorkersRequest
             {
